+++ Assets/OkapiKit/Scripts/Editor/ActionChangeSpriteRenderer.cs
@@ -38,16 +38,16 @@ namespace OkapiKit.Editor
                 EditorGUILayout.PropertyField(propTarget, new GUIContent("Target", "What's the sprite renderer to modify?"));
                 EditorGUILayout.PropertyField(propChangeType, new GUIContent("Change Type", "What kind of change to apply?\nSprite: Changes the image of the sprite renderer\nColor: Change the color of the sprite renderer.\nNote that if there's an animator linked to this object, the change can be overwritten by it."));
 
-                if (propChangeType.enumValueIndex == (int)ActionChangeSpriteRenderer.ChangeType.Sprite)
+                if (propChangeType.intValue == (int)ActionChangeSpriteRenderer.ChangeType.Sprite)
                 {
                     EditorGUILayout.PropertyField(propSprite, new GUIContent("Sprite", "What's the new sprite to set?"));
                 }
-                if (propChangeType.enumValueIndex == (int)ActionChangeSpriteRenderer.ChangeType.Color)
+                if (propChangeType.intValue == (int)ActionChangeSpriteRenderer.ChangeType.Color)
                 {
                     EditorGUILayout.PropertyField(propColor, new GUIContent("Color", "What's the new color we want to set?"));
                 }
-                if ((propChangeType.enumValueIndex == (int)ActionChangeSpriteRenderer.ChangeType.FlipX) ||
-                    (propChangeType.enumValueIndex == (int)ActionChangeSpriteRenderer.ChangeType.FlipY))
+                if ((propChangeType.intValue == (int)ActionChangeSpriteRenderer.ChangeType.FlipX) ||
+                    (propChangeType.intValue == (int)ActionChangeSpriteRenderer.ChangeType.FlipY))
                 {
                     EditorGUILayout.PropertyField(propBoolState, new GUIContent("Flip?", "Should we flip this axis?"));
                 }