+++ Assets/OkapiKit/Scripts/Editor/ActionSpawnEditor.cs
@@ -44,16 +44,16 @@ namespace OkapiKit.Editor
                     if (propPrefabObject.objectReferenceValue != null)
                     {
                         EditorGUILayout.PropertyField(propSpawnPosition, new GUIContent("Position", "What is the position of the new object?\nDefault: Object will spawn on the position where it was when the prefab was created\nThis: Object will spawn with the same position/rotation/scale as the current object\nTarget: Object will spawn at the position/rotation/scale of the target object\nTag: Object will spawn at the position/rotation/scale of an object tagged with the tag given."));
-                        if (propSpawnPosition.enumValueIndex == (int)ActionSpawn.SpawnPosition.Target)
+                        if (propSpawnPosition.intValue == (int)ActionSpawn.SpawnPosition.Target)
                         {
                             EditorGUILayout.PropertyField(propTargetPosition, new GUIContent("Target", "What's the target object where to spawn?"));
                         }
-                        else if (propSpawnPosition.enumValueIndex == (int)ActionSpawn.SpawnPosition.Tag)
+                        else if (propSpawnPosition.intValue == (int)ActionSpawn.SpawnPosition.Tag)
                         {
                             EditorGUILayout.PropertyField(propTargetTag, new GUIContent("Tag", "Find all objects with this tag, and select a random one where to spawn the new object."));
                         }
                     }
-                    if (propSpawnPosition.enumValueIndex != (int)ActionSpawn.SpawnPosition.Default)
+                    if (propSpawnPosition.intValue != (int)ActionSpawn.SpawnPosition.Default)
                     {
                         EditorGUILayout.PropertyField(propSetParent, new GUIContent("Set Parent", "Should we set the parent of the new object to this object?"));
                     }